feat: lots of refactoring and renaming files + documentation + timetable data management in settings
ps: documentation is stupid i think because i never did this before
