+++ lib/components/settings/screens/timetable_period_screen.dart
@@ -3,6 +3,7 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:timetable/constants/custom_times.dart';
 import 'package:timetable/models/settings.dart';
 
+/// 
 class TimetablePeriodScreen extends ConsumerWidget {
   const TimetablePeriodScreen({super.key});
