+++ lib/components/settings/timetable_features.dart
@@ -1,0 +1,50 @@
+import 'package:flutter/material.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:timetable/models/settings.dart';
+import 'package:timetable/components/settings/screens/timetable_period_screen.dart';
+
+/// All the settings for changing some timetable features.
+class TimetableFeaturesOptions extends ConsumerWidget {
+  const TimetableFeaturesOptions({super.key});
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final settings = ref.read(settingsProvider.notifier);
+    final rotationWeeks = ref.watch(settingsProvider).rotationWeeks;
+    final autoCompleteColor = ref.watch(settingsProvider).autoCompleteColor;
+
+    return Column(
+      children: [
+        ListTile(
+          onTap: () {
+            Navigator.push(
+              context,
+              MaterialPageRoute(
+                settings: const RouteSettings(
+                  name: "CellScreen",
+                ),
+                builder: (context) => const TimetablePeriodScreen(),
+              ),
+            );
+          },
+          title: const Text("Time Period Config"),
+        ),
+        SwitchListTile(
+          title: const Text("Rotation Weeks"),
+          subtitle: const Text("Experimental"),
+          value: rotationWeeks,
+          onChanged: (bool value) {
+            settings.updateRotationWeeks(value);
+          },
+        ),
+        SwitchListTile(
+          title: const Text("AutoComplete Colors"),
+          value: autoCompleteColor,
+          onChanged: (bool value) {
+            settings.updateAutoCompleteColor(value);
+          },
+        ),
+      ],
+    );
+  }
+}
