+++ lib/components/subject_management/subject_configs/day_config.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:timetable/components/widgets/bottom_sheets/days_modal_bottom_sheet.dart';
 import 'package:timetable/constants/days.dart';
 
+/// Day config part of the Subject creation screen.
 class DayConfig extends StatelessWidget {
   final ValueNotifier<Days> day;
