+++ lib/components/subject_management/subject_configs/note_tile.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:timetable/components/widgets/list_tile_group.dart';
 
+/// Note config part of the Subject creation screen.
 class NotesTile extends ConsumerWidget {
   final ValueNotifier<String?> note;
