+++ lib/components/widgets/bottom_sheets/rotation_week_modal_bottom_sheet.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:timetable/constants/rotation_weeks.dart';
 
+/// Bottom Sheet Modal Widget used to select a subject's rotation week.
 class RotationWeekModalBottomSheet extends StatelessWidget {
   final List<RotationWeeks> rotationWeeks;
   final ValueNotifier<RotationWeeks> rotationWeek;
