+++ lib/components/widgets/days_row.dart
@@ -1,9 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:timetable/constants/days.dart';
 import 'package:timetable/constants/grid_properties.dart';
 import 'package:timetable/models/settings.dart';
-import 'package:timetable/utilities/grid_utils.dart';
 
+/// Widget that appears at the top of the timetable grid view screen and shows the week's days.
 class DaysRow extends ConsumerWidget {
   const DaysRow({super.key});
