+++ lib/components/widgets/time_column.dart
@@ -1,9 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:timetable/constants/grid_properties.dart';
+import 'package:timetable/constants/time.dart';
 import 'package:timetable/models/settings.dart';
-import 'package:timetable/utilities/grid_utils.dart';
 
+/// Widget that appears at the left side of the timetable grid view screen and shows the timetable time period.
 class TimeColumn extends ConsumerWidget {
   const TimeColumn({super.key});
