+++ lib/constants/custom_times.dart
@@ -2,6 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:timetable/models/settings.dart';
 
+/// Returns the custom start time set by the user if customTimePeriod is true,
+/// otherwise returns the default start time. (8:00)
 TimeOfDay getCustomStartTime(TimeOfDay customTime, WidgetRef ref) {
   final customTimePeriod = ref.watch(settingsProvider).customTimePeriod;
 
@@ -16,6 +18,8 @@ TimeOfDay getCustomStartTime(TimeOfDay customTime, WidgetRef ref) {
   }
 }
 
+/// Returns the custom end time set by the user if customTimePeriod is true,
+/// otherwise returns the default end time. (18:00)
 TimeOfDay getCustomEndTime(TimeOfDay customTime, WidgetRef ref) {
   final customTimePeriod = ref.watch(settingsProvider).customTimePeriod;
 
@@ -30,6 +34,7 @@ TimeOfDay getCustomEndTime(TimeOfDay customTime, WidgetRef ref) {
   }
 }
 
+/// Returns the hour part of the customTime, formatted.
 String getCustomTimeHour(TimeOfDay customTime) {
   if (customTime.hour < 10) {
     return "0${customTime.hour}";
@@ -38,6 +43,7 @@ String getCustomTimeHour(TimeOfDay customTime) {
   }
 }
 
+/// Returns the minute part of the customTime, formatted.
 String getCustomTimeMinute(TimeOfDay customTime) {
   if (customTime.minute < 10) {
     return "0${customTime.minute}";
