+++ lib/constants/days.dart
@@ -1,3 +1,4 @@
+/// The week's days.
 enum Days {
   monday,
   tuesday,
@@ -7,3 +8,14 @@ enum Days {
   saturday,
   sunday,
 }
+
+/// The week's days as a list of strings.
+const List<String> days = [
+  'Monday',
+  'Tuesday',
+  'Wednesday',
+  'Thursday',
+  'Friday',
+  'Saturday',
+  'Sunday'
+];
