+++ lib/constants/theme_options.dart
@@ -1,3 +1,4 @@
+/// Theme mode options.
 enum ThemeModeOption {
   dark,
   light,
